[frontend] improve the readability of connector names (#3154)
